[#72] 코디 완료 후 홈 화면 새로고침 및 전환 애니메이션 적용
